fix(proxmox-sys): cast statfs.f_type to i64 for musl ABI compatibility
patches/proxmox-sys assigned libc statfs.f_type directly into an i64 field. That is glibc signed __fsword_t, but on musl (the x86_64-unknown-linux-musl build target) f_type is unsigned c_ulong, so the build failed with mismatched types (expected i64, found u64). f_type is an opaque filesystem magic identifier — no signed arithmetic is performed on it — so an explicit cast restores the intended i64 field on both ABIs (and on 32-bit, where glibc f_type is i32) without changing semantics.

X-StageFreight-Generated: true

+++ patches/proxmox-sys/src/fs/mod.rs
@@ -171,7 +171,10 @@ pub fn fs_info<P: ?Sized + nix::NixPath>(path: &P) -> nix::Result<FileSystemInfo
         available: stat.f_bavail * block_size,
         total_inodes: stat.f_files,
         free_inodes: stat.f_ffree,
-        fs_type: stat.f_type,
+        // f_type is glibc's signed __fsword_t (i64) but musl's unsigned c_ulong
+        // (u64); cast for the musl target. The value is an opaque fs magic id, so
+        // the reinterpretation is harmless — no signed arithmetic is done on it.
+        fs_type: stat.f_type as i64,
         fs_id: stat.f_fsid,
     })
 }
